+++ RELEASE_NOTES.md
@@ -12,6 +12,7 @@
   - UUIDv7 codec (16 bytes instead of 36 bytes string)
   - ULID codec (16 bytes instead of 26 bytes string)
 - `LMDBKodec` renamed to `KeyCodec` and moved to the `keycodecs` module
+- cross-collections transaction using `lift` through several collections 
 
 ## 2.2 - 2025-01-30